+++ contracts/test/MockToken.sol
@@ -4,13 +4,14 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract MockToken is ERC20("Mock", "MCT") {
     uint8 _decimals;
+
     constructor(uint8 __decimals) {
         _decimals = __decimals;
     }
 
-    function setBalanceTo(address to, uint value) public {
-		_mint(to, value);
-	}
+    function setBalanceTo(address to, uint256 value) public {
+        _mint(to, value);
+    }
 
     function mint(address _to, uint256 _amount) external {
         _mint(_to, _amount);
@@ -20,7 +21,7 @@ contract MockToken is ERC20("Mock", "MCT") {
         _burn(_from, _amount);
     }
 
-    function decimals() public view override returns(uint8) {
+    function decimals() public view override returns (uint8) {
         return _decimals;
     }
 }
